test(worker): split Consistency tests into their own file
cloud/workers/tests/test_analyze_basic.py crossed the 1200-line hard
cap for test files after adding the 3 new Consistency tests. Extract
them into cloud/workers/tests/test_reliability_consistency.py with
their own module-level preamble and a dedicated test class.

After the split:
  test_analyze_basic.py:           1164 lines (was 1331)
  test_reliability_consistency.py:  202 lines (new)

All 3 extracted tests continue to pass locally.
